this is for you dad
added accessibility toggle for people with epilepsy or other conditions that prevents them from viewing strobing, flashing or combined lights for long periods of time.  Some flashing is left as the game sequences relies on those for progression such as light sequence changes..  some sprites will also flash, regardless if the sector light is disabled or not.