Vanilla accuracy changes - part 1
When dying, artifacts are cleared, as well as any backpacks, weapons and ammo.  This is N64 behaviour, I consider this a bug / cheating in EX and it is banished without mercy!  Also added back a clause in seg rendering that was not there for some reason, angles are pretty important to get 'right'.