A huge cleanup of types
The typedefs are gone and we just use stdint.h whenever we need.  Only three typedefs remain and those are the ones that are critical to 64bit compatibility with shootable switches and dehacked usage.  I've gone back to using unsigned long / signed long in some places to preserve the 32bit compatibility where it is needed.  We should not use fixed width for these as for Linux/Mac and Unix, a long is still 64bit.  Only on Windows is it 32bit and so a fixed width integer would actually break the code on Unix as it would be expecting a 64bit type, on windows it expects a 32bit type.  So a long is an easy fix.