Removing the C89 from the library and doing a change
Ok, this and the last commit gets the linux/mac/bsd builds booting up again.  However they now fail with a P_LoadLeafs: vertex out of range: 947 - 569 error.

The issue with the previous way was that traditionally, dboolean was an integer, not a bool.  So changing this to a bool actually broke several areas where it was expected to have an int.

I took the code from Chocolate Doom which adds boolean as an unsigned char, allowing some form of 'cheating' for boolean expressions that actually handle characters, rather than boolean values or integer values.

Now we must go through the code and make it clean and use correct types for everything, I think this is the best way to get the code to be clean, rather than typedeffing boolean values.  It is more effort, but at least we will know that all the functions we change are using their correct type rather than a 'one size fits all' approach.

Windows works fine, but this is ok.  We had to break it in order to fix up all this mess, no issues from me.