Using Kaiser's version of fluidsynth for EX Classic
So there was always this random crash inside the fluid_synth_write_s16 with fluidlite and it was absolutely annoying as hell.  No matter what I did, it would randomly crash.  Kaiser's version is obviously based on a very old version of fluidsynth, but since moving to it, I have not had a single crash in 60 minutes of gameplay.  It also doesn't 'crackle' as much.  SOLD!