Fixing some things
So it turns out (for whatever reason) EX+ didn't like some duplicated frames for the spider mastermind and would (sometimes) display a warning about memory not being read when exiting the game.  I've removed a few frames and modified it so it still displays the entire death animation and it appears to have fixed the issue.

Other change is for 32bit fluidsynth library and dll