Skies are fixed
Shaders improved too.  As for the 'actual sky'  I think this is a remaster bug where the colours are actually bugged, this was discovered about 2 years ago and obviously can't be fixed on their side.  I found it out when porting over reloaded where marsh would be orange instead of green, because of the values in the SKYDEF.  Let's try to change those values to 'actual known' colours and that'll probably work.  Skies are clearly there, just the wrong colour.