Adjusting the amplitude of the factor of light manipulation
This adjusts the 'flickers' in the lights.  Adjusted to a value that mimics the N64, where flickering lights are naturally brighter as these sector specials also need to be brighter to influence the environment, like the final room in MAP01.  Too low of a value and it isn't visible, too high and it'll blow out.  With correct brightness and gamma settings for your monitor, these values are fine and mimic the N64 version pretty well.