Removing mouselook and fixing some alpha
I know some folks won't like this, but removing mouselook has to be done.

The non-skybox skies are not fit for mouselook as they are in screen-space and will follow the player's view.  I will not be doing skies anymore, I've focused on them for almost a month getting them like this.  Nightdive didn't have to deal with that as mouselook isn't in the remaster.  And this, as a hobby project shouldn't be expected to have my personal time wasted.